Add a test framework.
A test takes a formula, attempts to prove it, then runs it through
the Isabelle code generator to obtain an Isabelle file proving the
formula using SeCaV.
It then generates an Isabelle session and externally executes
the Isabelle build system to verify that the proof is correct.

Unfortunately, the test suite can currently not clean up after
itself, so the Makefile should be used to run tests for now.