Change the proof system to make proofs simpler.
The new proof system applies proof rules to all applicable formulas in
the sequent at once.
This is a drastic departure from the previous proof system.
The idea is that the new proof system should hopefully make proving
Hintikkaicity of the formulas along an escape path, and thus the
completeness proof, easier.

This change breaks the Haskell interface to the prover since there are
major differences in how proof states are represented and how proofs
should be post-processed into proper SeCaV proofs.
Proofs now require much more post-processing to become proper SeCaV
proofs, since the Ext rule must be inserted between most rule
applications, and rule applications that do nothing must be removed.
This could probably be achieved without too much trouble, but a verified
implementation seems like it would be quite annoying.

The non-Hintikkaicity part of the completeness proof works with small
changes to accommodate the new type of proof states.
A sketch of the Hintikkaicity proof is included for representative
connectives, but work is still needed.
A number of lemmas are still unproven (mostly about the relations
between proof states on an escape path).

Note also that there is an unfinished proof of the fairness of the cycle
of proof rules in the file Prover.thy.
This should be easy to prove, but involves some reasoning about infinite
cycles, and showing that they actually repeat.

The soundness proof is also still completely unfinished, but the hope is
that some of the lemmas from the Hintikkaicity proof may be reused.